refactor: move agent utilities to dedicated lib/agents module
Extracts `normalizeAgentRecord` from `lib/types.ts` into a new
`lib/agents.ts` module. Updates imports in /api/agents and
/api/leaderboard routes. Type definitions remain in lib/types.ts.

Closes #374
